Moved Inline input messages
Moved Inline input messages to InlineQuery/Type as that is a better
place for them. Also fixed recursiveness not being in
MySQLStorageHandlers load method.
